+++ app/uk/gov/hmrc/cdsreimbursementclaimfrontend/views/claims/enter_associated_mrn.scala.html
@@ -19,7 +19,7 @@
 @import play.twirl.api.Html
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.config.ViewConfig
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.actions.RequestWithSessionData
-@import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.claims.routes
+@import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsmultiple.{routes=>overpaymentsMultipleRoutes}
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.answers.AssociatedMrn
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.ids.AssociatedMrnIndex
 
@@ -41,8 +41,8 @@
 @hasErrors = @{form.hasErrors || form.hasGlobalErrors}
 
 @submitUrl = @{
-    if(editing) routes.EnterAssociatedMrnController.submitChangedMrn(index)
-    else routes.EnterAssociatedMrnController.submitEnteredMrn(index)
+    if(editing) overpaymentsMultipleRoutes.EnterAssociatedMrnController.submitChangedMrn(index)
+    else overpaymentsMultipleRoutes.EnterAssociatedMrnController.submitEnteredMrn(index)
 }
 
 @layout(pageTitle = Some(title), hasErrors = hasErrors) {
